split bootstrap interface startup
